+++ packages/react-app/src/components/Answer/AnswerForm.tsx
@@ -9,7 +9,6 @@ import {useContractFunction, useEthers} from "@usedapp/core";
 import {Contract} from "@ethersproject/contracts";
 import {RealityETH_v3_0__factory} from "../../typechain";
 import Alert from "@mui/material/Alert";
-import { hexZeroPad, hexlify } from "@ethersproject/bytes";
 import { BigNumber } from "@ethersproject/bignumber";
 import {Event} from "../../graphql/subgraph";
 import FormHelperText from "@mui/material/FormHelperText";
@@ -17,15 +16,12 @@ import {formatAmount, getAnswerText, getTimeLeft, isFinalized, showWalletError} 
 import CircularProgress from '@mui/material/CircularProgress';
 import { Trans, t } from "@lingui/macro";
 import {useI18nContext} from "../../lib/I18nContext";
-import {REALITY_TEMPLATE_MULTIPLE_SELECT} from "../../lib/reality";
+import {formatOutcome, REALITY_TEMPLATE_MULTIPLE_SELECT, INVALID_RESULT, ANSWERED_TOO_SOON} from "../../lib/reality";
 
-export const INVALID_RESULT = "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff";
-export const ANSWERED_TOO_SOON = "0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe";
-
-type ValidOutcomes = number | typeof INVALID_RESULT | typeof ANSWERED_TOO_SOON;
+export type FormEventOutcomeValue = number | typeof INVALID_RESULT | typeof ANSWERED_TOO_SOON;
 
 export type AnswerFormValues = {
-  outcome: ValidOutcomes | [ValidOutcomes] | ''
+  outcome: FormEventOutcomeValue | [FormEventOutcomeValue] | ''
 }
 
 type AnswerFormProps = {
@@ -54,7 +50,7 @@ export default function AnswerForm({event, register, errors, handleSubmit, setSh
     setCurrentBond(lastBond.gt(0) ? lastBond.mul(2) : minBond);
 
     // outcomes
-    let _outcomes: {value: string|number, text: string}[] = [];
+    let _outcomes: {value: FormEventOutcomeValue, text: string}[] = [];
 
     _outcomes = event.outcomes
         // first map and then filter to keep the index of each outcome as value
@@ -94,26 +90,9 @@ export default function AnswerForm({event, register, errors, handleSubmit, setSh
   }
 
   const onSubmit = async (data: AnswerFormValues) => {
-    let answer = ''
-
-    if (typeof data.outcome === 'object') {
-      // TODO: Update the multiple select if there are answers with invalid or too soon.
-      // this options are incompatible with multi-select
-      if (data.outcome.includes(INVALID_RESULT)) {
-        answer = INVALID_RESULT
-      } else if (data.outcome.includes(ANSWERED_TOO_SOON)) {
-        answer = ANSWERED_TOO_SOON
-      } else {
-        const answerChoice = (data.outcome as number[]).reduce((partialSum: number, value: number) => partialSum + 2 ** value, 0);
-        answer = hexZeroPad(hexlify(answerChoice), 32);
-      }
-    } else {
-      answer = hexZeroPad(hexlify(data.outcome), 32)
-    }
-
     await send(
       event.id,
-      answer,
+      formatOutcome(data.outcome),
       currentBond,
       {
         value: currentBond
